separate Gantt chart util functions into 2 separate files
