+++ app/components/GanttChart/GanttChartBehavior.tsx
@@ -3,56 +3,9 @@
 import { RefObject } from 'react';
 // date
 import { CalendarDate } from '@internationalized/date';
-// schemas
-import { Project, Sprint } from '@schemas';
 // styles
 import './frappe-gantt.css';
 import './frappe-gantt-custom.css';
-
-/********************************************************************************************************************
- * types
- ********************************************************************************************************************/
-export type GanttTask = {
-  id: string;
-  name: string;
-  start: string;  // yyy-mm-dd
-  end: string;    // yyy-mm-dd
-  progress: number;
-  custom_class?: string;
-};
-
-/********************************************************************************************************************
- * format sprints into Frappe Gantt-compatible structure
- ********************************************************************************************************************/
-export function formatSprintToGanttTask(sprint: Sprint): GanttTask {
-  return {
-    id: sprint.id,
-    name: sprint.title,
-    start: sprint.startDate.toString(),
-    end: sprint.dueDate.toString(),
-    progress: 70, // placeholder
-    custom_class: 'gantt-task-bar',
-  }
-}
-
-export function formatSprintsToGanttTasks(sprints: Project['sprints']): GanttTask[] {
-  return sprints.map((sprint) => formatSprintToGanttTask(sprint));
-}
-
-/********************************************************************************************************************
- * format a GanttTask back into Sprint
- ********************************************************************************************************************/
-export function formatGanttTaskToSprint(task: GanttTask, originalSprint: Sprint): Sprint {
-  const startParts = task.start.split('-').map(Number);
-  const endParts = task.end.split('-').map(Number);
-
-  return {
-    ...originalSprint,
-    title: task.name,
-    startDate: new CalendarDate(startParts[0], startParts[1], startParts[2]),
-    dueDate: new CalendarDate(endParts[0], endParts[1], endParts[2]),
-  };
-}
 
 /********************************************************************************************************************
  * injects a vertical deadline marker to Gantt chart
@@ -129,4 +82,22 @@ export function doCustomScroll(
       });
     }
   }
+}
+
+/********************************************************************************************************************
+ * disables horizontal scrolling via mouse wheel on a given element
+ * - still allows scrollbar dragging and vertical scroll
+ ********************************************************************************************************************/
+export function disableHorizontalWheelScroll(container: HTMLElement | null) {
+  const onWheel = (event: WheelEvent) => {
+    if (event.deltaX !== 0) {
+      event.preventDefault();
+    }
+  };
+
+  container?.addEventListener('wheel', onWheel, { passive: false });
+
+  return () => {
+    container?.removeEventListener('wheel', onWheel);
+  };
 }
